+++ src/extensions/map/mapext_hooks.cpp
@@ -1,0 +1,99 @@
+/*******************************************************************************
+/*                 O P E N  S O U R C E  --  V I N I F E R A                  **
+/*******************************************************************************
+ *
+ *  @project       Vinifera
+ *
+ *  @file          MAPEXT_HOOKS.CPP
+ *
+ *  @author        Rampastring
+ *
+ *  @brief         Contains the hooks for the extended MapClass.
+ *
+ *  @license       Vinifera is free software: you can redistribute it and/or
+ *                 modify it under the terms of the GNU General Public License
+ *                 as published by the Free Software Foundation, either version
+ *                 3 of the License, or (at your option) any later version.
+ *
+ *                 Vinifera is distributed in the hope that it will be
+ *                 useful, but WITHOUT ANY WARRANTY; without even the implied
+ *                 warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ *                 PURPOSE. See the GNU General Public License for more details.
+ *
+ *                 You should have received a copy of the GNU General Public
+ *                 License along with this program.
+ *                 If not, see <http://www.gnu.org/licenses/>.
+ *
+ ******************************************************************************/
+#include "mapext_hooks.h"
+#include "map.h"
+#include "mouseext_hooks.h"
+#include "mouse.h"
+#include "mousetype.h"
+#include "object.h"
+#include "objecttype.h"
+#include "vinifera_globals.h"
+#include "wwmouse.h"
+#include "debughandler.h"
+#include "asserthandler.h"
+#include "extension.h"
+#include "weapontype.h"
+#include "cell.h"
+
+#include "hooker.h"
+#include "hooker_macros.h"
+#include "techno.h"
+#include "tibsun_functions.h"
+#include "weapontypeext.h"
+
+
+ /**
+  *  A fake class for implementing new member functions which allow
+  *  access to the "this" pointer of the intended class.
+  *
+  *  @note: This must not contain a constructor or destructor!
+  *  @note: All functions must be prefixed with "_" to prevent accidental virtualization.
+  */
+static class MapClassExt final : public MapClass
+{
+public:
+    void _Place_Down(Cell& cell, ObjectClass* object);
+};
+
+
+/**
+ *  Fixes a buffer overflow crash in the original MapClass::Place_Down
+ *  where the game accessed the cell array before checking that the cell
+ *  number is within bounds.
+ *
+ *  @author: Rampastring
+ */
+void MapClassExt::_Place_Down(Cell& cell, ObjectClass* object)
+{
+    if (!object)
+        return;
+
+    if (object->Class_Of()->IsFootprint && object->In_Which_Layer() == LAYER_GROUND) {
+        Cell xlist[32];
+        List_Copy(object->Occupy_List(), std::size(xlist), xlist);
+        Cell const* list = xlist;
+        while (*list != REFRESH_EOL) {
+            Cell newcell = cell + *list++;
+            int cellnum = newcell.As_Cell_Number();
+
+            if (cellnum >= 0 && cellnum < MAP_CELL_TOTAL && Array[cellnum] != nullptr) {
+                (*this)[newcell].Occupy_Down(object, object->IsOnBridge);
+                (*this)[newcell].Recalc_Attributes();
+            }
+        }
+    }
+}
+
+
+/**
+ *  Main function for patching the hooks.
+ */
+void MapClassExtension_Hooks()
+{
+    Patch_Jump(0x00511070, &MapClassExt::_Place_Down);
+}