+++ docs/Whats-New.md
@@ -74,6 +74,7 @@ Vanilla fixes:
 - Fix the player being able to land an aircraft onto a helipad even if it's not listed as `Dock=` of this aircraft (by ZivDero)
 - Fix a bug where a weeder could be ordered to dock with a waste facility that wasn't listed in the weeder's `Dock=` key (by ZivDero)
 - Fix a bug where the player couldn't repair aircraft on allied repair bays under some circumstances (by ZivDero)
+- Fix a buffer overflow crash in MapClass::Place_Down (by Rampastring)
 
 :::